+++ qmcpy/kernel_methods/kernel/_kernel_prod.py
@@ -23,7 +23,7 @@ class _KernelProd(object):
         assert isinstance(scale,float)
         self.d = dimension
         if isinstance(lengthscales,float):
-            lengthscales = lengthscales*self.npt.ones(self.d,**self.ckwargs)
+            lengthscales = lengthscales*self.npt.ones(self.d,dtype=float,**self.ckwargs)
         self.lengthscales = lengthscales
         self.scale = scale
         if alpha is None: alpha = self.DEFAULTALPHA
@@ -55,7 +55,7 @@ class _KernelProd(object):
         m1 = len(beta1s) 
         m2 = len(beta2s)
         assert beta1s.shape==(m1,self.d) and beta2s.shape==(m2,self.d)
-        if isinstance(c1s,float): c1s = c1s*self.npt.ones(m1,**self.ckwargs)
-        if isinstance(c2s,float): c2s = c2s*self.npt.ones(m2,**self.ckwargs)
+        if isinstance(c1s,float): c1s = c1s*self.npt.ones(m1,dtype=float,**self.ckwargs)
+        if isinstance(c2s,float): c2s = c2s*self.npt.ones(m2,dtype=float,**self.ckwargs)
         assert c1s.shape==(m1,) and c2s.shape==(m2,)
         return self.parsed_call(x1,x2,n1,n2,beta1s,beta2s,m1,m2,c1s,c2s)
